Extract types from components in projectId page directory
